requirements.txt
blacksheep[full]==2.4.0
Jinja2~=3.1.6
uvloop; platform_system != "Windows"
essentials-configuration[full]
MarkupSafe==3.0.2
pydantic
aiosqlite==0.21.0
alembic==1.14.1
asyncpg==0.30.0
black==25.1.0
blacksheep-sqlalchemy==0.0.4
flake8==7.1.2
httptools==0.6.4
httpx==0.28.1
isort==6.0.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pytest-asyncio==0.25.3
pytest-cov==6.0.0
rich-click==1.8.8
uvicorn==0.34.0
